Make OdfConverter.cs build with Mono (using the Makefiles) again. Introduce a couple more #if !MONO conditionals.
Ideally, we should not have conditionally compiled code in any source
file here, but instead only do run-time check, for two unrelated
things: 1) if running on Mono or .NET, and 2) when running on Mono, if
on Linux (or Unix in general) or on Windows. Currently this all is a
bit mixed up, and we assume that Mono-compiled odf-converter code will
only run on Mono and .NET-compiled code will only run on .NET and thus
on Windows.

Ideally it should also be possible to use the Makefiles but still use
csc.exe and not Mono's gmcs. But fixing this all is probably not worth
the effort, at least not for me.