Compiling C# scripts embedded in XSL does not really work with just a statically bundled Mono runtime in the odf-converter executable, which is the context where the mkbundled odf-converter is intended to work.
So, we extract the C# code from those .xsl files and precompile it,
and bundle the resulting assemblies too. The name of such a
precompiled class contains the hash value of the C# code snippet in
question. Mono's XSLT knows to look for such a class when trying to
use embedded C# code.

From Novell's odf-converter-3.0-Shell-OdfConverter-Makefile.diff.